Add the files to ignore in the docker image
and specify how to handle the newlines on different platforms,
when checking out shell scripts (for Unix or Windows).

Without this setting checked out in the repository, users that clone
the project over Windows will have problems running scripts into
Docker containers, because files are being copied on it with the
wrong line-ending (Unix use LF; Windows uses CRLF).